fontes2022_replication/requirements.txt
numpy>=1.24
scipy>=1.10
pyyaml>=6.0
matplotlib>=3.7
fluidfoam>=0.2
